requirements.txt
# Core RAG and LLM libraries
langchain==0.1.0
llama-index==0.9.30
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0
pypdf2==3.0.1
pdfplumber==0.10.0

# Ollama integration
ollama==0.1.7
httpx==0.25.2

# Frontend
streamlit==1.28.2

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4
tiktoken==0.5.2
requests==2.31.0

# Optional: For advanced text processing
spacy==3.7.2
nltk==3.8.1